BRS-1094: Moving subarea code to activity endpoint
